backend/pyproject.toml
[tool.poetry]
name = "analyzer"
version = "0.1.0"
description = ""
authors = ["Guilherme <guibarasuol99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.3"
fastapi = "^0.109.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.26.0"
pydantic-settings = {extras = ["dotenv"], version = "^2.1.0"}
celery = {extras = ["librabbitmq"], version = "^5.3.6"}
sqlalchemy = "^2.0.25"

[tool.poetry.dev-dependencies]
black = { extras = ["d"], version = "^22.12.0" }
pytest = "^7.4.4"
autoflake = "^1.6.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^2.20.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
extend-exclude = "analyzer/db/migrations"

[tool.isort]
profile = "black"
extend_skip = ["analyzer", "db", "migrations"]
src_paths = ["analyzer", "tests"]
combine_as_imports = true

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
exclude = ["backend/analyzer/db/migrations"]
